Update the count badge every 15 minutes
Although the package already hooks in multiple places to update the
count badge as soon as the content is changed, there are still
situations when the count badge doesn't reflect the true state. One of
the examples is when the computer goes to sleep for a night and then
when it awakes today is already next day, but org-agenda-dock doesn't
check that. With this timer it works correctly for situations like
this.